requirements.txt
beautifulsoup4==4.14.2
faiss-gpu==1.7.2
fitz==0.0.1.dev2
numpy==2.3.4
pymupdf4llm==0.1.8
requests==2.32.5
sentence_transformers==5.1.2
tiktoken==0.12.0
torch==2.9.0+cpu
tqdm==4.67.1
transformers==4.57.1
trafilatura==2.0.0